Initial import for Spring 2011 (iTrust v11)
git-svn-id: https://itrust.svn.sourceforge.net/svnroot/itrust@140 12e484f7-5e12-421d-bbfb-6194dbfb34fd

+++ iTrust/src/edu/ncsu/csc/itrust/beans/loaders/PrescriptionBeanLoader.java
@@ -32,7 +32,7 @@ public class PrescriptionBeanLoader implements BeanLoader<PrescriptionBean> {
 	public PrescriptionBean loadSingle(ResultSet rs) throws SQLException {
 		PrescriptionBean pres = new PrescriptionBean();
 		pres.setId(rs.getLong("ID"));
-		pres.setVisitID((long) rs.getInt("VisitID"));
+		pres.setVisitID(rs.getInt("VisitID"));
 		pres.setStartDateStr(new SimpleDateFormat("MM/dd/yyyy").format(new java.util.Date(rs.getDate(
 				"StartDate").getTime())));
 		pres.setEndDateStr(new SimpleDateFormat("MM/dd/yyyy").format(new java.util.Date(rs.getDate("EndDate")
